+++ src/features/venue/components/OpeningHoursStatus/getOpeningHoursStatus.ts
@@ -1,19 +1,6 @@
 import { addHours, set, isWithinInterval } from 'date-fns'
 
-export type OpeningHours = Partial<{
-  MONDAY: OpeningHour
-  TUESDAY: OpeningHour
-  WEDNESDAY: OpeningHour
-  THURSDAY: OpeningHour
-  FRIDAY: OpeningHour
-  SATURDAY: OpeningHour
-  SUNDAY: OpeningHour
-}>
-
-type OpeningHour = {
-  open: string
-  close: string
-}[]
+import { OpeningHours } from 'features/venue/types'
 
 type OpeningHoursStatusParams = {
   openingHours: OpeningHours
@@ -46,7 +33,10 @@ const getStateFromPeriods = (
 ): OpeningHoursStatus => {
   const period = periods[index]
   if (!period) {
-    return closeState()
+    return {
+      state: 'close',
+      text: 'Fermé',
+    }
   }
 
   if (period.isPassed(currentDate)) {
@@ -55,13 +45,22 @@ const getStateFromPeriods = (
 
   if (period.isOpen(currentDate)) {
     if (period.isClosingSoon(currentDate)) {
-      return closeSoonState(period.closeAt)
+      return {
+        state: 'close-soon',
+        text: `Ferme bientôt - ${formatDate(period.closeAt)}`,
+      }
     }
-    return openState(period.closeAt)
+    return {
+      state: 'open',
+      text: `Ouvert jusqu’à ${formatDate(period.closeAt)}`,
+    }
   }
 
   if (period.isOpeningSoon(currentDate)) {
-    return openSoonState(period.openAt)
+    return {
+      state: 'open-soon',
+      text: `Ouvre bientôt - ${formatDate(period.openAt)}`,
+    }
   }
 
   return getStateFromPeriods(periods, currentDate, index + 1)
@@ -104,26 +103,6 @@ const getDateFromOpeningHour = (currentDate: Date, openingHour: string): Date =>
     minutes: minute,
   })
 }
-
-const closeState = (): OpeningHoursStatus => ({
-  state: 'close',
-  text: 'Fermé',
-})
-
-const openState = (openDate: Date): OpeningHoursStatus => ({
-  state: 'open',
-  text: `Ouvert jusqu’à ${formatDate(openDate)}`,
-})
-
-const openSoonState = (openDate: Date): OpeningHoursStatus => ({
-  state: 'open-soon',
-  text: `Ouvre bientôt - ${formatDate(openDate)}`,
-})
-
-const closeSoonState = (closeDate: Date): OpeningHoursStatus => ({
-  state: 'close-soon',
-  text: `Ferme bientôt - ${formatDate(closeDate)}`,
-})
 
 const formatDate = (date: Date): string => {
   const hours = date.getHours()
