+++ src/features/venue/types.ts
@@ -22,3 +22,18 @@ export type TabArrowNavigationProps = {
   setSelectedTab: (tab: Tab) => void
   tabs: Tab[]
 }
+
+export type OpeningHours = Partial<{
+  MONDAY: OpeningHour
+  TUESDAY: OpeningHour
+  WEDNESDAY: OpeningHour
+  THURSDAY: OpeningHour
+  FRIDAY: OpeningHour
+  SATURDAY: OpeningHour
+  SUNDAY: OpeningHour
+}>
+
+type OpeningHour = {
+  open: string
+  close: string
+}[]
